Move auth components to dedicated folder
