+++ src/components/auth/register/RegisterPage.tsx
