+++ src/pages/change-password.tsx
@@ -1,6 +1,6 @@
 import { GetStaticProps } from 'next'
 import { serverSideTranslations } from 'common/useNextLocale'
-import Page from 'components/changePassword/ChangePasswordPage'
+import Page from 'components/auth/changePassword/ChangePasswordPage'
 
 export const getStaticProps: GetStaticProps = async ({ locale }) => ({
   props: {
