+++ src/pages/login.tsx
@@ -1,5 +1,5 @@
 import { GetStaticProps } from 'next'
-import LoginPage from 'components/login/LoginPage'
+import LoginPage from 'components/auth/login/LoginPage'
 import { serverSideTranslations } from 'common/useNextLocale'
 
 export const getStaticProps: GetStaticProps = async ({ locale }) => ({
