+++ src/pages/register.tsx
@@ -1,5 +1,5 @@
 import { GetStaticProps } from 'next'
-import RegisterPage from 'components/register/RegisterPage'
+import RegisterPage from 'components/auth/register/RegisterPage'
 import { serverSideTranslations } from 'common/useNextLocale'
 
 export const getStaticProps: GetStaticProps = async ({ locale }) => ({
